% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backProp.R
\name{backProp}
\alias{backProp}
\title{Title}
\usage{
backProp(
  x,
  y,
  weightMatrix,
  activation,
  reluLeak,
  modelType,
  eta,
  gradientClip,
  baisUnits,
  regularisePar,
  itr,
  optimiser,
  parMomentum,
  parRmsProp,
  parRmsPropZeroAdjust,
  previousWeightUpdate,
  previousBiasUpdate,
  previousWeightAdapt,
  previousBiasAdapt,
  inputSizeImpact
)
}
\arguments{
\item{inputSizeImpact}{}
}
\value{

}
\description{
Title
}
