% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepForest.R
\name{deepforest}
\alias{deepforest}
\title{Title}
\usage{
deepforest(
  x,
  y,
  networkCount = 3,
  layerChoice = c(2:3),
  unitsChoice = c(4:10),
  cutVarSizePercent = 0.6,
  cutDataSizePercent = 0.6,
  activation = "relu",
  reluLeak = 0,
  modelType = "regress",
  iterations = 500,
  eta = 10^-2,
  seed = 2,
  gradientClip = 0.8,
  regularisePar = 0,
  optimiser = "adam",
  parMomentum = 0.9,
  inputSizeImpact = 1,
  parRmsPropZeroAdjust = 10^-8,
  parRmsProp = 0.9999,
  treeLeaves = NA,
  treeMinSplitPercent = 0.3,
  treeMinSplitCount = 100,
  treeCp = 0.01,
  errorCover = 0.2,
  treeAugment = T,
  printItrSize = 100,
  showProgress = T,
  stopError = 0.01,
  miniBatchSize = NA,
  useBatchProgress = T
)
}
\arguments{
\item{treeAugment}{}
}
\value{

}
\description{
Title
}
\examples{

}
