% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedForward.R
\name{feedForward}
\alias{feedForward}
\title{Title}
\usage{
feedForward(xBiased, weightMatrix, activation, reluLeak, modelType, baisUnits)
}
\arguments{
\item{baisUnits}{}
}
\value{

}
\description{
Title
}
