% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniBatchCreate.R
\name{miniBatchCreate}
\alias{miniBatchCreate}
\title{Mini Batch Creator for deepnet}
\usage{
miniBatchCreate(xrows, miniBatchSize, prevBatch)
}
\arguments{
\item{prevBatch}{}
}
\value{

}
\description{
Mini Batch Creator for deepnet
}
