% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiActivationList.R
\name{multiActivationList}
\alias{multiActivationList}
\title{Activation List for Deepforest}
\usage{
multiActivationList(multiLayerList, activation, seed)
}
\arguments{
\item{activation}{}
}
\value{

}
\description{
Activation List for Deepforest
}
