% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDataCut.R
\name{multiDataCut}
\alias{multiDataCut}
\title{Random Data Cut for deepforest}
\usage{
multiDataCut(x, y, networkCount, cutDataSizePercent, cutVarSizePercent, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Random Data Cut for deepforest
}
