% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLayerIntialiser.R
\name{multiLayerIntialiser}
\alias{multiLayerIntialiser}
\title{Title}
\usage{
multiLayerIntialiser(layerChoice, unitsChoice, networkCount, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Title
}
