% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVariableCut.R
\name{multiVariableCut}
\alias{multiVariableCut}
\title{Random Variable cuts for deepforest}
\usage{
multiVariableCut(x, cutVarSizePercent, networkCount, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Random Variable cuts for deepforest
}
