% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiserAdjust.R
\name{optimiserAdjust}
\alias{optimiserAdjust}
\title{OptimserAdjustments to Gradent for deepnet}
\usage{
optimiserAdjust(
  i,
  dw,
  db,
  optimiser,
  parMomentum,
  parRmsProp,
  parRmsPropZeroAdjust,
  previousWeightUpdate,
  previousBiasUpdate,
  previousWeightAdapt,
  previousBiasAdapt
)
}
\arguments{
\item{previousBiasAdapt}{}
}
\value{

}
\description{
OptimserAdjustments to Gradent for deepnet
}
