% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictdeepForest.R
\name{predict.deepforest}
\alias{predict.deepforest}
\title{Predict Function for DeepForest}
\usage{
\method{predict}{deepforest}(object, newData)
}
\arguments{
\item{newData}{}
}
\value{

}
\description{
Predict Function for DeepForest
}
