% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictdeepnet.R
\name{predict.deepnet}
\alias{predict.deepnet}
\title{Predict Function for Deepnet}
\usage{
\method{predict}{deepnet}(object, newData)
}
\arguments{
\item{newData}{}
}
\value{

}
\description{
Predict Function for Deepnet
}
