% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictdeeptree.R
\name{predict.deeptree}
\alias{predict.deeptree}
\title{Title}
\usage{
\method{predict}{deeptree}(object, newData, treeLeaves = NA, stackPred = NA)
}
\arguments{
\item{stackPred}{}
}
\value{

}
\description{
Title
}
