% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffler.R
\name{shuffler}
\alias{shuffler}
\title{shuffler for varialble importance}
\usage{
shuffler(x, colname, y, model, modelError, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
shuffler for varialble importance
}
