% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{variableImportance}
\alias{variableImportance}
\title{Variable importance for models in this library}
\usage{
variableImportance(model, x, y, showPlot = T, seed = 2)
}
\arguments{
\item{model}{Model object}

\item{x}{a data frame with input variables}

\item{y}{a data frame with ouptut variable}

\item{showPlot}{logical. True will show importance plot. Default True}

\item{seed}{Set seed with this parameter. Incase of sin activation sometimes changing seed can yeild better results. Default is 2}
}
\value{
returns variable importance data frame
}
\description{
Variable importance for models in this library
}
