% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightUpdate.R
\name{weightUpdate}
\alias{weightUpdate}
\title{Title}
\usage{
weightUpdate(
  i,
  AllWeights,
  feedOut,
  x,
  y,
  ypred,
  zin,
  activation,
  reluLeak,
  eta,
  gradientClip,
  regularisePar,
  itr,
  optimiser,
  parMomentum,
  parRmsProp,
  parRmsPropZeroAdjust,
  sizeImpact
)
}
\arguments{
\item{sizeImpact}{}
}
\value{

}
\description{
Title
}
