% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightUpdateInitialiser.R
\name{weightUpdateInitialiser}
\alias{weightUpdateInitialiser}
\title{Weight Update Initialisation for deepnet}
\usage{
weightUpdateInitialiser(i, interVariableCount, interOutCount, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Weight Update Initialisation for deepnet
}
