% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLayerIntialiser.R
\name{multiLayerIntialiser}
\alias{multiLayerIntialiser}
\title{Multi Layer Initialisation for deepforest}
\usage{
multiLayerIntialiser(layerChoice, unitsChoice, networkCount, seed)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Multi Layer Initialisation for deepforest
}
\examples{

}
