% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.deepforest.R
\name{predict.deepforest}
\alias{predict.deepforest}
\title{Predict Function for DeepForest}
\usage{
\method{predict}{deepforest}(object, newData, ...)
}
\arguments{
\item{object}{deepforest model object}

\item{newData}{pass dataframe for prediction}

\item{...}{further arguments passed to or from other methods.}
}
\value{
returns predictions vector or dataframe
}
\description{
Predict Function for DeepForest
}
