% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.deepnet.R
\name{predict.deepnet}
\alias{predict.deepnet}
\title{Predict Function for Deepnet}
\usage{
\method{predict}{deepnet}(object, newData, ...)
}
\arguments{
\item{object}{deepnet model object}

\item{newData}{pass dataframe for prediction}

\item{...}{further arguments passed to or from other methods.}
}
\value{
returns predictions vector or dataframe
}
\description{
Predict Function for Deepnet
}
