% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.deeptree.R
\name{predict.deeptree}
\alias{predict.deeptree}
\title{Predict Function for Deeptree}
\usage{
\method{predict}{deeptree}(object, newData, treeLeaves = NA, stackPred = NA, ...)
}
\arguments{
\item{object}{deeptree model object}

\item{newData}{pass dataframe for prediction}

\item{treeLeaves}{Pass vector with tree leaves if fit outside deeptree. default NA.}

\item{stackPred}{Pass stackPred of prediction data if it was passed in deeptree}

\item{...}{further arguments passed to or from other methods.}
}
\value{
returns predictions vector or dataframe
}
\description{
Predict Function for Deeptree
}
