% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightIntialiser.R
\name{weightInitialiser}
\alias{weightInitialiser}
\title{Title}
\usage{
weightInitialiser(i, interVariableCount, interOutCount, seed, activation)
}
\arguments{
\item{activation}{}
}
\value{

}
\description{
Title
}
